cambios en la creación de alquiler, los precios ahora tienen un valor random de entre 3 y 6, cambios en el listado de alquileres filtrados (se aplicara tambien a peliculas) donde se aplican los controles para mostrar la lista en relación a si hay orden y filtros o no.

+++ src/main/java/com/movies4rent/Servidor/DTO/CreaAlquilerDTO.java
@@ -18,7 +18,7 @@ public class CreaAlquilerDTO {
     private LocalDate fechaFin;
     private EstadoAlquiler estado;
 
-    private double precio;
+    private Integer precio;
     public LocalDate getFechaInicio() {
         return fechaInicio;
     }
@@ -44,11 +44,11 @@ public class CreaAlquilerDTO {
     }
 
 
-    public double getPrecio() {
+    public Integer getPrecio() {
         return precio;
     }
 
-    public void setPrecio(double precio) {
+    public void setPrecio(Integer precio) {
         this.precio = precio;
     }
